ps6 uploaded
Dear Akito, I am aware of the fact that this assignment is one day late. However, circumstances have not been the ideal for me. At the moment the person that I interacted with has confirmed that he doesn’t have the virus (which is a relief). Nevertheless, the possibility of having it altered my last 4 days as I had to stay in quarantine waiting for the test results. I hope you can make an exception, I really try to submit my assignments on time but this events made it very complicated to even think about homework.